end of the lesson 27.02.2025

+++ 05_week/toDoRestApp-backEnd/src/controller/todos.ts
@@ -1,13 +1,14 @@
 import { RequestHandler } from "express";
+import mongoose from "mongoose";
 import Todo, { TodoModel } from "../model/todo";
 
 export const createTodo: RequestHandler = async (req, res, next) => {
 	try {
 		const data: TodoModel = req.body;
-		let todos = await Todo.create(data);
-		return res.status(200).json({ message: "Todo created successfully", data: todos });
+		await Todo.create(data);
+		res.status(200).json({ message: "Todo created successfully", data });
 	} catch (error: any) {
-		return res.status(500).json({ message: error.message });
+		res.status(500).json({ message: error.message });
 	}
 };
 
@@ -33,6 +34,9 @@ export const updateTodo: RequestHandler = async (req, res, next) => {
 export const deleteTodo: RequestHandler = async (req, res, next) => {
 	try {
 		const { id } = req.params;
+		if (!mongoose.Types.ObjectId.isValid(id)) {
+			return res.status(400).json({ message: "Invalid ID format" });
+		}
 		let isDeleted = await Todo.findByIdAndDelete(id);
 		if (!isDeleted) throw new Error("Failed to delete todo!");
 		return res.status(200).json({ message: "Todo deleted successfully!" });
